+++ src/main/kotlin/team/mozu/dsm/application/port/in/lesson/command/UpdateLessonCommand.kt
@@ -2,6 +2,6 @@ package team.mozu.dsm.application.port.`in`.lesson.command
 
 data class UpdateLessonCommand(
     val lessonName: String,
-    val baseMoney: Int,
+    val baseMoney: Long,
     val lessonRound: Int
 )
