+++ src/main/kotlin/team/mozu/dsm/domain/lesson/model/LessonItem.kt
@@ -7,17 +7,17 @@ import team.mozu.dsm.domain.lesson.model.id.LessonItemId
 @Aggregate
 data class LessonItem(
     val lessonItemId: LessonItemId,
-    val currentMoney: Int,
-    val round1Money: Int,
-    val round2Money: Int,
-    val round3Money: Int,
-    val round4Money: Int?,
-    val round5Money: Int?
+    val currentMoney: Long,
+    val round1Money: Long,
+    val round2Money: Long,
+    val round3Money: Long,
+    val round4Money: Long?,
+    val round5Money: Long?
 ) {
     /**
      * 특정 차수의 가격 조회
      */
-    fun getPriceByRound(round: Int): Int? {
+    fun getPriceByRound(round: Int): Long? {
         return when (round) {
             0 -> currentMoney
             1 -> round1Money
